Restructure into src/scgist package, split utility.py, scanpy to benchmarks group (milestone 3)
