--- scGIST/__init__.py
@@ -1,3 +1,0 @@
-# __init__.py
-from .scGIST import scGIST
-from .utility import *
