+++ src/scgist/__init__.py
@@ -1,0 +1,13 @@
+from scgist.model import scGIST
+from scgist.evaluation import test_classifier
+from scgist.priority import get_priority_score_list
+from scgist.plotting import plot_confusion_matrix, plot_marker_weights, plot_history
+
+__all__ = [
+    "scGIST",
+    "test_classifier",
+    "get_priority_score_list",
+    "plot_confusion_matrix",
+    "plot_marker_weights",
+    "plot_history",
+]
